a lot of code refactoring
