+++ components/__init__.py
@@ -1,7 +1,82 @@
-from .navbar import NavBar
-from .uploadbutton import UploadButton
-from .alerts import ErrorAlert, FilenameAlert
-from .header import Header
-from .plotplaceholder import PlotPlaceHolder
-from .cards import UploadCard, ContactUploadCard, DisplayControlCard, WarningsCard, HelpCard
-from .modals import MismatchModal, MissingInput_Modal
+def MissingInput_Modal(*args, **kwargs):
+    from components.modals import MissingInput_Modal
+
+    return MissingInput_Modal(*args, **kwargs)
+
+
+def MismatchModal(*args, **kwargs):
+    from components.modals import MismatchModal
+
+    return MismatchModal(*args, **kwargs)
+
+
+def HelpCard(*args, **kwargs):
+    from components.cards import HelpCard
+
+    return HelpCard(*args, **kwargs)
+
+
+def WarningsCard(*args, **kwargs):
+    from components.cards import WarningsCard
+
+    return WarningsCard(*args, **kwargs)
+
+
+def ContactUploadCard(*args, **kwargs):
+    from components.cards import ContactUploadCard
+
+    return ContactUploadCard(*args, **kwargs)
+
+
+def ContactDisplayControlCard(*args, **kwargs):
+    from components.cards import DisplayControlCard
+
+    return ContactDisplayControlCard(*args, **kwargs)
+
+
+def NavBar(*args, **kwargs):
+    from components.navbar import NavBar
+
+    return NavBar(*args, **kwargs)
+
+
+def Header(*args, **kwargs):
+    from components.header import Header
+
+    return Header(*args, **kwargs)
+
+
+def UploadButton(*args, **kwargs):
+    from components.uploadbutton import UploadButton
+
+    return UploadButton(*args, **kwargs)
+
+
+def FilenameAlert(*args, **kwargs):
+    from components.alerts import FilenameAlert
+
+    return FilenameAlert(*args, **kwargs)
+
+
+def ErrorAlert(*args, **kwargs):
+    from components.alerts import ErrorAlert
+
+    return ErrorAlert(*args, **kwargs)
+
+
+def PlotPlaceHolder(*args, **kwargs):
+    from components.plotplaceholder import PlotPlaceHolder
+
+    return PlotPlaceHolder(*args, **kwargs)
+
+
+def UploadCard(*args, **kwargs):
+    from components.cards import UploadCard
+
+    return UploadCard(*args, **kwargs)
+
+
+def DisplayControlCard(*args, **kwargs):
+    from components.cards import DisplayControlCard
+
+    return DisplayControlCard(*args, **kwargs)
