+++ components/alerts.py
@@ -1,6 +1,6 @@
 import dash_bootstrap_components as dbc
 import dash_html_components as html
-from index import PathIndex
+from core import PathIndex
 
 
 def ErrorAlert(is_open=False):
