--- index/__init__.py
@@ -1,3 +1,0 @@
-from .pathindex import PathIndex
-from .componentindex import *
-from .statesindex import *
