+++ parsers/topconsparser.py
@@ -1,5 +1,5 @@
-from .parser import Parser
-from index.statesindex import MembraneStates
+from parsers import MembraneStates
+from parsers.parser import Parser
 
 
 class TopconsParser(Parser):
